Create and style base-state sidebar.
Created three divs within the sidebar. The first one for the title and
logo, and the resting two for the options the user can use, which
consist of a simple image followed by a descriptive word. All the divs
have been positioned using flexbox and/or margins.

The sidebar also has a fourth element. An img which is position: absolute
relative to the sidebar, and shows an arrow icon which the user will use
to minimize the sidebar.