Style interactive elements and add retract-expand functionality at sidebar.
The divs that contain the different sidebar options are given a dark
green background-color and solid border-left when hovered.

The sidebar-handler "img-button" has been renamed to "sidebar-handler-
retract", since it will be used to minimize the sidebar.
A second flipped img-button called "sidebar-handler-expand" has been
created and positioned where the sidebar remains when it is retracted.
This second img-button is unseen by default and will be used to expand
the sidebar when the conditions are met.

Both buttons are given a rotation animation on hover.

A retract-expand functionality has been added to the sidebar element
through CSS and JS:
- Function minimizeSidebar (later to be renamed) contains two event
triggers.
- When clicking the retract img-button, the header and main-content element are
applied properties that change their base positioning on the grid,
invading the second column and leaving the text part of the sidebar
content unseen, while the icons remain visible. The expand button
becomes visible as well.
- When clicking the expand img-button, all the properties previously
applied are removed, so the layout returns to the base state.