Create and provide basic styling for main-content section
Main-content section is divided in two different columns with grid. It
is also given a lower z-index than the sidebar section, so that when it
occupies its space it is displayed over it.

The first column is the biggest and contains a title and below a grid of
divs (each one is used to display a project.). This grid uses the repeat
and auto-fit functions to distribute the grid across arrow or different
columns depending on the available width in the section.

--> Each project div contains a title, an image that links to the project's
live web, a short explanation of the project and three small buttons to
like the project, view it's live web, or view it's github page.

The second column is divided in four rows with grid. The first and the
third are for the titles and second and fourth for two divs that contain
some anouncements in the first and user profiles that are trending in the
second.

--> Each of the divs have its title change colour on hover simulating
links.