Add hover and click funcionality for the project icons
Created "styleIcons" function on the js file. It targets all the projects
icons and reacts on the following events:
--> On mouseenter, the three icon outlines turn to green (the original
img src is substituted by another one).
--> On mouseleave, the three icons go back to the original image with
the black outline, except the heart icon given it contains the class
"clicked".
--> On click, only the heart icon is affected. It turns and is completely
filled up with the green colour, and adds the class "clicked". If the
user clicks the heart when it already has the class "clicked" (when it
is green), it turns again to the normal colour and removes the class
"clicked".

Links were added to some elements and href attributes completed for
the project icon links.