Add Dark/Light Mode functionality
Created a Dark/Light mode functionality triggered by the checkbox
(toggle):

--> Colours for each mode are now stored in variables. The ones for
lightmode are based in the :root element and the ones for the dark on
the .dark class, which is add using JavaScript to the Body element.

--> The dark mode uses is applied by adding the .dark class to the Body
element when the toggle is clicked. It is also created in the localStorage
so that when the user leaves the page on this mode, it will be kept the
next time it is visited.

--> Images need to be changed of colour in antoher way. New images with
the wanted colour have been added to the img folder to be used for the
dark mode. They are stored in variable and added as a new src in the place
they belong, whenever body has the .dark class applied. Light mode images
are also re-added, to make sure user can switch back dark mode and the
images do so too.

Some missing links have also been added to the header section.